DroidGuard: Clean up and deduplicate
