+++ play-services-droidguard/core/src/main/java/com/google/android/gms/droidguard/DroidGuardChimeraService.java
@@ -17,7 +17,7 @@ import com.google.android.gms.framework.tracing.wrapper.TracingIntentService;
 
 import org.microg.gms.droidguard.core.DroidGuardServiceBroker;
 import org.microg.gms.droidguard.GuardCallback;
-import org.microg.gms.droidguard.core.HandleProxyFactory;
+import org.microg.gms.droidguard.core.NetworkHandleProxyFactory;
 import org.microg.gms.droidguard.PingData;
 import org.microg.gms.droidguard.Request;
 
@@ -30,7 +30,7 @@ import java.util.concurrent.TimeUnit;
 public class DroidGuardChimeraService extends TracingIntentService {
     public static final Object a = new Object();
     // factory
-    public HandleProxyFactory b;
+    public NetworkHandleProxyFactory b;
     // widevine
     public Object c;
     // executor
@@ -51,7 +51,7 @@ public class DroidGuardChimeraService extends TracingIntentService {
         setIntentRedelivery(true);
     }
 
-    public DroidGuardChimeraService(HandleProxyFactory factory, Object ping, Object database) {
+    public DroidGuardChimeraService(NetworkHandleProxyFactory factory, Object ping, Object database) {
         super("DG");
         setIntentRedelivery(true);
         this.b = factory;
@@ -120,7 +120,7 @@ public class DroidGuardChimeraService extends TracingIntentService {
     @Override
     public void onCreate() {
         this.e = new Object();
-        this.b = new HandleProxyFactory(this);
+        this.b = new NetworkHandleProxyFactory(this);
         this.g = new Object();
         this.h = new Handler();
         this.c = new Object();
