+++ play-services-droidguard/core/src/main/kotlin/org/microg/gms/droidguard/core/NetworkHandleProxyFactory.kt
@@ -6,13 +6,11 @@
 package org.microg.gms.droidguard.core
 
 import android.content.Context
-import androidx.annotation.GuardedBy
 import com.android.volley.NetworkResponse
 import com.android.volley.VolleyError
 import com.android.volley.toolbox.RequestFuture
 import com.android.volley.toolbox.Volley
 import com.google.android.gms.droidguard.internal.DroidGuardResultsRequest
-import dalvik.system.DexClassLoader
 import okio.ByteString.Companion.decodeHex
 import okio.ByteString.Companion.of
 import org.microg.gms.droidguard.*
@@ -20,16 +18,11 @@ import org.microg.gms.profile.Build
 import org.microg.gms.profile.ProfileManager
 import org.microg.gms.utils.singleInstanceOf
 import java.io.File
-import java.io.IOException
-import java.security.MessageDigest
-import java.security.cert.Certificate
 import java.util.*
 import com.android.volley.Request as VolleyRequest
 import com.android.volley.Response as VolleyResponse
 
-class HandleProxyFactory(private val context: Context) {
-    @GuardedBy("CLASS_LOCK")
-    private val classMap = hashMapOf<String, Class<*>>()
+class NetworkHandleProxyFactory(private val context: Context) : HandleProxyFactory(context) {
     private val dgDb: DgDatabaseHelper = DgDatabaseHelper(context)
     private val version = VersionUtil(context)
     private val queue = singleInstanceOf { Volley.newRequestQueue(context.applicationContext) }
@@ -136,7 +129,7 @@ class HandleProxyFactory(private val context: Context) {
 
             override fun getHeaders(): Map<String, String> {
                 return mapOf(
-                        "User-Agent" to "DroidGuard/${version.versionCode}"
+                    "User-Agent" to "DroidGuard/${version.versionCode}"
                 )
             }
         })
@@ -178,68 +171,7 @@ class HandleProxyFactory(private val context: Context) {
         return HandleProxy(clazz, context, flow, byteCode, callback, vmKey, extra, request?.bundle)
     }
 
-    fun getTheApkFile(vmKey: String) = File(getCacheDir(vmKey), "the.apk")
-    private fun getCacheDir() = context.getDir(CACHE_FOLDER_NAME, Context.MODE_PRIVATE)
-    private fun getCacheDir(vmKey: String) = File(getCacheDir(), vmKey)
-    private fun getOptDir(vmKey: String) = File(getCacheDir(vmKey), "opt")
-    private fun isValidCache(vmKey: String) = getTheApkFile(vmKey).isFile && getOptDir(vmKey).isDirectory
-
-    private fun updateCacheTimestamp(vmKey: String) {
-        try {
-            val timestampFile = File(getCacheDir(vmKey), "t")
-            if (!timestampFile.exists() && !timestampFile.createNewFile()) {
-                throw Exception("Failed to touch last-used file for $vmKey.")
-            }
-            if (!timestampFile.setLastModified(System.currentTimeMillis())) {
-                throw Exception("Failed to update last-used timestamp for $vmKey.")
-            }
-        } catch (e: IOException) {
-            throw Exception("Failed to touch last-used file for $vmKey.")
-        }
-    }
-
-    private fun verifyApkSignature(apk: File): Boolean {
-        return true
-        val certificates: Array<Certificate> = TODO()
-        if (certificates.size != 1) return false
-        return Arrays.equals(MessageDigest.getInstance("SHA-256").digest(certificates[0].encoded), PROD_CERT_HASH)
-    }
-
-    private fun loadClass(vmKey: String, bytes: ByteArray): Class<*> {
-        synchronized(CLASS_LOCK) {
-            val cachedClass = classMap[vmKey]
-            if (cachedClass != null) {
-                updateCacheTimestamp(vmKey)
-                return cachedClass
-            }
-            val weakClass = weakClassMap[vmKey]
-            if (weakClass != null) {
-                classMap[vmKey] = weakClass
-                updateCacheTimestamp(vmKey)
-                return weakClass
-            }
-            if (!isValidCache(vmKey)) {
-                throw BytesException(bytes, "VM key $vmKey not found in cache")
-            }
-            if (!verifyApkSignature(getTheApkFile(vmKey))) {
-                getCacheDir(vmKey).deleteRecursively()
-                throw ClassNotFoundException("APK signature verification failed")
-            }
-            val loader = DexClassLoader(getTheApkFile(vmKey).absolutePath, getOptDir(vmKey).absolutePath, null, context.classLoader)
-            val clazz = loader.loadClass(CLASS_NAME)
-            classMap[vmKey] = clazz
-            weakClassMap[vmKey] = clazz
-            return clazz
-        }
-    }
-
     companion object {
-        const val CLASS_NAME = "com.google.ccc.abuse.droidguard.DroidGuard"
         const val SERVER_URL = "https://www.googleapis.com/androidantiabuse/v1/x/create?alt=PROTO&key=AIzaSyBofcZsgLSS7BOnBjZPEkk4rYwzOIz-lTI"
-        const val CACHE_FOLDER_NAME = "cache_dg"
-        val CLASS_LOCK = Object()
-        @GuardedBy("CLASS_LOCK")
-        val weakClassMap = WeakHashMap<String, Class<*>>()
-        val PROD_CERT_HASH = byteArrayOf(61, 122, 18, 35, 1, -102, -93, -99, -98, -96, -29, 67, 106, -73, -64, -119, 107, -5, 79, -74, 121, -12, -34, 95, -25, -62, 63, 50, 108, -113, -103, 74)
     }
 }
