+++ play-services-droidguard/src/main/kotlin/org/microg/gms/droidguard/HandleProxy.kt
@@ -6,21 +6,39 @@
 package org.microg.gms.droidguard
 
 import android.content.Context
+import android.os.Bundle
 import android.os.Parcelable
 
-class HandleProxy(private val handle: Any, val vmKey: String) {
-    constructor(clazz: Class<*>, context: Context, vmKey: String, data: Parcelable) : this(kotlin.runCatching {
-        clazz.getDeclaredConstructor(Context::class.java, Parcelable::class.java).newInstance(context, data)
-    }.getOrElse {
-        throw it
-    }, vmKey
+class HandleProxy(val handle: Any, val vmKey: String, val extra: ByteArray = ByteArray(0)) {
+    constructor(clazz: Class<*>, context: Context, vmKey: String, data: Parcelable) : this(
+        kotlin.runCatching {
+            clazz.getDeclaredConstructor(Context::class.java, Parcelable::class.java).newInstance(context, data)
+        }.getOrElse {
+            throw BytesException(ByteArray(0), it)
+        },
+        vmKey
     )
+
+    constructor(clazz: Class<*>, context: Context, flow: String?, byteCode: ByteArray, callback: Any, vmKey: String, extra: ByteArray, bundle: Bundle?) : this(
+        kotlin.runCatching {
+            clazz.getDeclaredConstructor(Context::class.java, String::class.java, ByteArray::class.java, Object::class.java, Bundle::class.java).newInstance(context, flow, byteCode, callback, bundle)
+        }.getOrElse {
+            throw BytesException(extra, it)
+        }, vmKey, extra)
+
+    fun run(data: Map<Any, Any>): ByteArray {
+        try {
+            return handle.javaClass.getDeclaredMethod("run", Map::class.java).invoke(handle, data) as ByteArray
+        } catch (e: Exception) {
+            throw BytesException(extra, e)
+        }
+    }
 
     fun init(): Boolean {
         try {
             return handle.javaClass.getDeclaredMethod("init").invoke(handle) as Boolean
         } catch (e: Exception) {
-            throw e
+            throw BytesException(extra, e)
         }
     }
 
@@ -28,7 +46,8 @@ class HandleProxy(private val handle: Any, val vmKey: String) {
         try {
             handle.javaClass.getDeclaredMethod("close").invoke(handle)
         } catch (e: Exception) {
-            throw e
+            throw BytesException(extra, e)
         }
     }
+
 }
