+++ play-services-droidguard/src/main/kotlin/org/microg/gms/droidguard/HandleProxyFactory.kt
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2022 microG Project Team
+ * SPDX-FileCopyrightText: 2021 microG Project Team
  * SPDX-License-Identifier: Apache-2.0
  */
 
@@ -8,19 +8,56 @@ package org.microg.gms.droidguard
 import android.content.Context
 import android.os.Bundle
 import android.os.ParcelFileDescriptor
+import android.os.Parcelable
+import androidx.annotation.GuardedBy
 import dalvik.system.DexClassLoader
 import java.io.File
 import java.io.IOException
-import java.util.UUID
+import java.security.MessageDigest
+import java.security.cert.Certificate
+import java.util.*
 
-class HandleProxyFactory(private val context: Context) {
-    private fun getTheApkFile(vmKey: String) = File(getCacheDir(vmKey), "the.apk")
-    private fun getCacheDir() = context.getDir(CACHE_FOLDER_NAME, Context.MODE_PRIVATE)
-    private fun getCacheDir(vmKey: String) = File(getCacheDir(), vmKey)
-    private fun getOptDir(vmKey: String) = File(getCacheDir(vmKey), "opt")
-    private fun isValidCache(vmKey: String) = getTheApkFile(vmKey).isFile && getOptDir(vmKey).isDirectory
+open class HandleProxyFactory(private val context: Context) {
+    @GuardedBy("CLASS_LOCK")
+    protected val classMap = hashMapOf<String, Class<*>>()
 
-    private fun updateCacheTimestamp(vmKey: String) {
+    fun createHandle(vmKey: String, pfd: ParcelFileDescriptor, extras: Bundle): HandleProxy {
+        fetchFromFileDescriptor(pfd, vmKey)
+        return createHandleProxy(vmKey, extras)
+    }
+
+    private fun fetchFromFileDescriptor(pfd: ParcelFileDescriptor, vmKey: String) {
+        if (!isValidCache(vmKey)) {
+            val auIs = ParcelFileDescriptor.AutoCloseInputStream(pfd)
+            val temp = File(getCacheDir(), "${UUID.randomUUID()}.apk")
+            temp.parentFile!!.mkdirs()
+            temp.writeBytes(auIs.readBytes())
+            auIs.close()
+            getOptDir(vmKey).mkdirs()
+            temp.renameTo(getTheApkFile(vmKey))
+            updateCacheTimestamp(vmKey)
+            if (!isValidCache(vmKey)) {
+                getCacheDir(vmKey).deleteRecursively()
+                throw IllegalStateException("Error ")
+            }
+        }
+    }
+
+    private fun createHandleProxy(
+        vmKey: String,
+        extras: Parcelable
+    ): HandleProxy {
+        val clazz = loadClass(vmKey)
+        return HandleProxy(clazz, context, vmKey, extras)
+    }
+
+    fun getTheApkFile(vmKey: String) = File(getCacheDir(vmKey), "the.apk")
+    protected fun getCacheDir() = context.getDir(CACHE_FOLDER_NAME, Context.MODE_PRIVATE)
+    protected fun getCacheDir(vmKey: String) = File(getCacheDir(), vmKey)
+    protected fun getOptDir(vmKey: String) = File(getCacheDir(vmKey), "opt")
+    protected fun isValidCache(vmKey: String) = getTheApkFile(vmKey).isFile && getOptDir(vmKey).isDirectory
+
+    protected fun updateCacheTimestamp(vmKey: String) {
         try {
             val timestampFile = File(getCacheDir(vmKey), "t")
             if (!timestampFile.exists() && !timestampFile.createNewFile()) {
@@ -36,49 +73,35 @@ class HandleProxyFactory(private val context: Context) {
 
     private fun verifyApkSignature(apk: File): Boolean {
         return true
+        val certificates: Array<Certificate> = TODO()
+        if (certificates.size != 1) return false
+        return Arrays.equals(MessageDigest.getInstance("SHA-256").digest(certificates[0].encoded), PROD_CERT_HASH)
     }
 
-    private fun copyTheApk(pfd: ParcelFileDescriptor, vmKey: String) {
-        if (!isValidCache(vmKey)) {
-            val auIs = ParcelFileDescriptor.AutoCloseInputStream(pfd)
-            val temp = File(getCacheDir(), "${UUID.randomUUID()}.apk")
-            temp.parentFile!!.mkdirs()
-            temp.writeBytes(auIs.readBytes())
-            auIs.close()
-            getOptDir(vmKey).mkdirs()
-            temp.renameTo(getTheApkFile(vmKey))
-            updateCacheTimestamp(vmKey)
-            if (!isValidCache(vmKey)) {
-                getCacheDir(vmKey).deleteRecursively()
-                throw IllegalStateException("unknown except")
+    protected fun loadClass(vmKey: String, bytes: ByteArray = ByteArray(0)): Class<*> {
+        synchronized(CLASS_LOCK) {
+            val cachedClass = classMap[vmKey]
+            if (cachedClass != null) {
+                updateCacheTimestamp(vmKey)
+                return cachedClass
             }
-        }
-    }
-
-    fun createHandle(vmKey: String, pfd: ParcelFileDescriptor, extras: Bundle): HandleProxy {
-        copyTheApk(pfd, vmKey)
-        val clazz = loadClass(vmKey)
-        return HandleProxy(clazz, context, vmKey, extras)
-    }
-
-    private fun loadClass(vmKey: String): Class<*> {
-        val clazz = classMap[vmKey]
-        if (clazz != null) {
-            updateCacheTimestamp(vmKey)
-            return clazz
-        } else {
+            val weakClass = weakClassMap[vmKey]
+            if (weakClass != null) {
+                classMap[vmKey] = weakClass
+                updateCacheTimestamp(vmKey)
+                return weakClass
+            }
             if (!isValidCache(vmKey)) {
-                throw RuntimeException("VM key $vmKey not found in cache")
+                throw BytesException(bytes, "VM key $vmKey not found in cache")
             }
             if (!verifyApkSignature(getTheApkFile(vmKey))) {
                 getCacheDir(vmKey).deleteRecursively()
                 throw ClassNotFoundException("APK signature verification failed")
             }
-            val loader = DexClassLoader(
-                getTheApkFile(vmKey).absolutePath, getOptDir(vmKey).absolutePath, null, context.classLoader
-            )
+            val loader = DexClassLoader(getTheApkFile(vmKey).absolutePath, getOptDir(vmKey).absolutePath, null, context.classLoader)
             val clazz = loader.loadClass(CLASS_NAME)
             classMap[vmKey] = clazz
+            weakClassMap[vmKey] = clazz
             return clazz
         }
     }
@@ -86,6 +109,9 @@ class HandleProxyFactory(private val context: Context) {
     companion object {
         const val CLASS_NAME = "com.google.ccc.abuse.droidguard.DroidGuard"
         const val CACHE_FOLDER_NAME = "cache_dg"
-        val classMap = hashMapOf<String, Class<*>>()
+        val CLASS_LOCK = Object()
+        @GuardedBy("CLASS_LOCK")
+        val weakClassMap = WeakHashMap<String, Class<*>>()
+        val PROD_CERT_HASH = byteArrayOf(61, 122, 18, 35, 1, -102, -93, -99, -98, -96, -29, 67, 106, -73, -64, -119, 107, -5, 79, -74, 121, -12, -34, 95, -25, -62, 63, 50, 108, -113, -103, 74)
     }
 }
